tests/requirements.txt
boto3==1.26.32
PyYAML==6.0